+++ R/data_descriptions.R
@@ -1,4 +1,4 @@
-#' ANZSCO
+#' ANZSCO 2009
 #'
 #' Wide table containing all levels of the Australian and New Zealand Standard
 #' Classification of Occupations (ANZSCO), First Edition, Revision 1, 2009. Cat. 1220.0.
@@ -19,6 +19,54 @@
 #' See \url{https://www.abs.gov.au/ausstats/abs@.nsf/Previousproducts/C4BECE1704987586CA257089001A9181 } for details.}
 #' }
 "anzsco2009"
+
+
+#' ANZSCO 2019
+#'
+#' Wide table containing all levels of the Australian and New Zealand Standard
+#' Classification of Occupations (ANZSCO), Version 1.3, 2019
+#'
+#' @format A \code{tibble} with 11 variables:
+#' \describe{
+#' \item{\code{anzsco_major_code}}{Major ANZSCO group codes, e.g. "4"}
+#' \item{\code{anzsco_major}}{Major ANZSCO group names in title case, e.g. "Community and Personal Service Workers"}
+#' \item{\code{anzsco_submajor_code}}{Sub-major ANZSCO group codes, e.g. "45"}
+#' \item{\code{anzsco_submajor}}{Sub-major ANZSCO group names in title case, e.g. "Sports and Personal Service Workers"}
+#' \item{\code{anzsco_minor_code}}{Minor ANZSCO group codes, e.g. "452"}
+#' \item{\code{anzsco_minor}}{Minor ANZSCO group names in title case, e.g. "Sports and Fitness Workers"}
+#' \item{\code{anzsco_unit_code}}{Unit ANZSCO group codes, e.g. "4523"}
+#' \item{\code{anzsco_unit}}{Unit ANZSCO group names in title case, e.g. "Sports Coaches, Instructors and Officials"}
+#' \item{\code{anzsco_occupation_code}}{Occupation ANZSCO group codes, e.g. "452311"}
+#' \item{\code{anzsco_occupation}}{Occupation ANZSCO group names in title case, e.g. "Diving Instructor (Open Water)"}
+#' \item{\code{skill_level}}{Skill level required for occupation, determined by the ABS (1 is highest, 5 is lowest).
+#' See \url{https://www.abs.gov.au/ausstats/abs@.nsf/Previousproducts/C4BECE1704987586CA257089001A9181 } for details.}
+#' }
+"anzsco2019"
+
+
+
+#' ANZSCO 2013
+#'
+#' Wide table containing all levels of the Australian and New Zealand Standard
+#' Classification of Occupations (ANZSCO), Version 1.2, 2013
+#'
+#' @format A \code{tibble} with 11 variables:
+#' \describe{
+#' \item{\code{anzsco_major_code}}{Major ANZSCO group codes, e.g. "4"}
+#' \item{\code{anzsco_major}}{Major ANZSCO group names in title case, e.g. "Community and Personal Service Workers"}
+#' \item{\code{anzsco_submajor_code}}{Sub-major ANZSCO group codes, e.g. "45"}
+#' \item{\code{anzsco_submajor}}{Sub-major ANZSCO group names in title case, e.g. "Sports and Personal Service Workers"}
+#' \item{\code{anzsco_minor_code}}{Minor ANZSCO group codes, e.g. "452"}
+#' \item{\code{anzsco_minor}}{Minor ANZSCO group names in title case, e.g. "Sports and Fitness Workers"}
+#' \item{\code{anzsco_unit_code}}{Unit ANZSCO group codes, e.g. "4523"}
+#' \item{\code{anzsco_unit}}{Unit ANZSCO group names in title case, e.g. "Sports Coaches, Instructors and Officials"}
+#' \item{\code{anzsco_occupation_code}}{Occupation ANZSCO group codes, e.g. "452311"}
+#' \item{\code{anzsco_occupation}}{Occupation ANZSCO group names in title case, e.g. "Diving Instructor (Open Water)"}
+#' \item{\code{skill_level}}{Skill level required for occupation, determined by the ABS (1 is highest, 5 is lowest).
+#' See \url{https://www.abs.gov.au/ausstats/abs@.nsf/Previousproducts/C4BECE1704987586CA257089001A9181 } for details.}
+#' }
+"anzsco2013"
+
 
 
 
@@ -223,3 +271,13 @@
 #' }
 #'
 "asc_technology_tools_ranking"
+
+
+#' palette_state_name_2016
+#'
+#' Named vector with the 'unofficial colours of each State and Territory in Australia.
+#'
+#' @format A named \code{vector} with hexadecimal colours.
+
+"palette_state_name_2016"
+