+++ data-raw/create_anzsic2006.R
@@ -10,13 +10,8 @@ include_factor_variants <- FALSE
 # ty asiripanich
 anzsic_url <- "https://raw.githubusercontent.com/asiripanich/anzsic/master/anzsic_2006.csv"
 
-temp_dir <- tempdir()
-temp_path <- glue("{temp_dir}/anzsic.csv")
-
-download.file(anzsic_url, temp_path)
-
 # Read
-anzsic_raw <- read_csv(temp_path) %>%
+anzsic_raw <- read_csv(anzsic_url) %>%
     rename_all(~ glue("anzsic_{.}")) %>%
     mutate_if(is.double, as.integer) %>%
     as_tibble()
@@ -59,7 +54,7 @@ if (include_factor_variants) {
     mutate(across(ends_with("title"), fct_inorder, .names = "{.col}_f"))
 }
 
-anzsic2006 <- anzsic %>%
+anzsic2006 <- anzsic2006  %>%
   rename_with(~ str_remove_all(., "\\_title"), everything()) %>%
   arrange(anzsic_division_code,
           anzsic_subdivision_code,